Reduce complexity of reservation impl
We extract the date validation into its own class
