+++ src/reservation/reservation.impl.ts
@@ -1,5 +1,6 @@
 import dayjs from "dayjs";
 import {Reservation} from "./reservation";
+import {isValidDate} from "./validator.date";
 
 export class Identifiable {
     id: string | null | undefined;
@@ -14,7 +15,7 @@ export class ReservationImpl extends Identifiable implements Reservation {
 
     constructor(at: Date|string, email: string, name: string | null, quantity: number) {
         super();
-        if (!at || !ReservationImpl.isValidDate(at)) {
+        if (!at || !isValidDate(at)) {
             throw new Error(`${at} is not a valid date`);
         }
         if (!email || email === '') {
@@ -29,27 +30,5 @@ export class ReservationImpl extends Identifiable implements Reservation {
         this.email = email;
         this.name = name ?? "";
         this.quantity = quantity;
-    }
-
-
-    static isValidDate(date: Date|string) {
-        function findDateByPattern(myString: string) {
-            const yyyyMMdd = /\d{4}-\d{2}-\d{2}/;
-            return yyyyMMdd.test(myString);
-        }
-        let extractedDate;
-        if (date instanceof Date) {
-            extractedDate = new Date(date).toISOString();
-        } else if (findDateByPattern(date)) {
-            extractedDate = date;
-        } else if (/[12][0129]\d\d/.test(date)) {
-            extractedDate = new Date(date).toISOString();
-        } else {
-            return false;
-        }
-        const localDate = extractedDate.substring(0, "YYYY-MM-DD".length);
-        const dayjs1 = dayjs(localDate, "YYYY-MM-DD");
-        const s = dayjs1.format("YYYY-MM-DD");
-        return s === localDate;
     }
 }
